Turned project into a library, more functionality
The functionality of this project and the virtual machine has now been
completed, perhaps argument handling in the test program could be improved, but
the VM itself has all the needed functionality.

The project has now also been turned into a crate, which can be used in other
programs. It may need reworking, but this will become apparent only after I
learn Rust more.

Documentation for the library has been added, but there are no proper tests.
Examples in the documentation have been added, which compile!